Random Meme Generator/requirements.txt
Pillow==8.3.1
QuoteEngine==0.1.0
docx==0.2.4
pandas==1.3.3
Flask==2.0.1
requests==2.26.0